Full firewall rules are being loaded, not functional yet due to wildcard

+++ aya-workspace/firewhal-kernel-common/src/lib.rs
@@ -29,8 +29,8 @@ pub enum Action {
 #[derive(Clone, Copy, Hash, Eq, PartialEq)]
 pub struct RuleKey {
     pub protocol: u32,
-    pub source_port: u32,
-    pub dest_port: u32,
+    pub source_port: u16,
+    pub dest_port: u16,
     pub source_ip: u32,
     pub dest_ip: u32,
 }
